requirements.txt
google-cloud-datastore
google-cloud-storage
sqlalchemy
flask
ConfigParser
Jinja2
click
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
python-slugify